Submission_Task_2/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==23.2.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cog==0.13.6
contourpy @ file:///croot/contourpy_1732540045555/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
fastapi==0.110.3
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fonttools @ file:///croot/fonttools_1713551344105/work
fsspec==2024.10.0
ftfy==6.3.1
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.26.3
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1730902924303/work
kiwisolver @ file:///work/perseverance-python-buildout/croot/kiwisolver_1698847502605/work
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
networkx @ file:///croot/networkx_1720002482208/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
open_clip_torch==2.29.0
packaging @ file:///croot/packaging_1720101850331/work
pillow @ file:///croot/pillow_1731594689713/work
ply @ file:///work/perseverance-python-buildout/croot/ply_1698846188635/work
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing @ file:///croot/pyparsing_1731445506121/work
PyQt5==5.15.10
PyQt5-sip @ file:///work/perseverance-python-buildout/croot/pyqt-split_1698847927472/work/pyqt_sip
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.0.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.1.0
sip @ file:///work/perseverance-python-buildout/croot/sip_1698846977033/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
starlette==0.37.2
structlog==24.4.0
sympy==1.13.1
timm==1.0.12
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.67.1
transformers==4.47.0
triton==3.1.0
typing_extensions==4.12.2
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1727769808118/work
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
wheel==0.44.0
